Corrected View Settings issue.
Corrected ViewSettings issue to ensure data is shared for an entire
request across controls and partial views.  There was an issue with the
previous ViewData extension method when partial views were used.

Also corrected display of error message that occurs when a grid is first
loaded on a page.  Previously the wrong message box was being used, if
an edit window existed on the page.